+++ src/api/comm.c
@@ -1,22 +1,11 @@
-#include "api.h"
+#include "comm.h"
+#include "api_helper.h"
 #include "ipc/cryptCommunicator.h"
 #include "utils/logger.h"
 #include "utils/oidc_error.h"
 
 #include <stdarg.h>
 #include <stdlib.h>
-
-#ifndef API_LOGLEVEL
-#define API_LOGLEVEL NOTICE
-#endif  // API_LOGLEVEL
-
-#ifndef START_APILOGLEVEL
-#define START_APILOGLEVEL int oldLogMask = logger_setloglevel(API_LOGLEVEL);
-#endif
-#ifndef END_APILOGLEVEL
-#define END_APILOGLEVEL logger_setlogmask(oldLogMask);
-#endif  // END_APILOGLEVEL
-
 
 char* communicate(unsigned char remote, const char* fmt, ...) {
   START_APILOGLEVEL
@@ -32,7 +21,3 @@ char* communicate(unsigned char remote, const char* fmt, ...) {
   END_APILOGLEVEL
   return ret;
 }
-
-char* oidcagent_serror() { return oidc_serror(); }
-
-void oidcagent_perror() { oidc_perror(); }
