+++ src/api/comm.h
@@ -1,0 +1,11 @@
+#ifndef OIDC_API_GENERAL_H
+#define OIDC_API_GENERAL_H
+
+#include "export_symbols.h"
+
+char* communicate(unsigned char remote, const char* fmt, ...);
+
+
+
+
+#endif  // OIDC_API_GENERAL_H
