+++ src/api/error.c
@@ -1,0 +1,7 @@
+#include "error.h"
+
+#include "utils/oidc_error.h"
+
+char* oidcagent_serror() { return oidc_serror(); }
+
+void oidcagent_perror() { oidc_perror(); }
