--- src/oidc-add/parse.h
@@ -1,8 +1,0 @@
-#ifndef OIDC_ADD_PARSE_H
-#define OIDC_ADD_PARSE_H
-
-#include "api.h"
-
-struct loaded_accounts_response parseForLoadedAccountsListResponse(char* response);
-
-#endif /* OIDC_ADD_PARSE_H */
