--- src/oidc-token/parse.h
@@ -1,8 +1,0 @@
-#ifndef OIDC_TOKEN_PARSE_H
-#define OIDC_TOKEN_PARSE_H
-
-#include "api.h"
-
-struct token_response parseForTokenResponse(char* response);
-
-#endif /* OIDC_TOKEN_PARSE_H */
